move `migrate-arbitrary-utilities` to core
Note: this commit on its own _will_ have some failing tests because the
`migrate-arbitrary-utilities.test.ts` file combined some other
migrations as well.

Eventually once those are ported over, will make the tests pass.

Another thing to notice is that we updated some tests. The more we
canonicalize candidates, the more tests will need upgrading to an even
more canonical form.
